refactor: mv filter_metrics to prepare_data.py from heatmaps.py
